training/requirements.txt
# DeepSeek OCR Training Requirements (Unsloth)
# =============================================
# Unsloth provides 1.4x faster training with 40% less VRAM

# Unsloth (RECOMMENDED)
unsloth
unsloth_zoo

# PyTorch and CUDA support
torch>=2.0.0
torchvision>=0.15.0

# Transformers and related
transformers>=4.45.0
tokenizers>=0.15.0
accelerate>=0.25.0
datasets>=2.16.0

# Trainer
trl>=0.9.0

# Parameter-efficient fine-tuning
peft>=0.7.0

# Image processing
Pillow>=10.0.0

# Data processing
numpy>=1.24.0
tqdm>=4.66.0

# Logging
tensorboard>=2.15.0

# Quantization
bitsandbytes>=0.41.0

# Evaluation
jiwer>=3.0.0
rapidfuzz>=3.0.0
